fix comment and re-structure lib architecture
